backend/requirements.txt
Flask
flask-cors
PyMuPDF
pytesseract
Pillow
nltk
sumy
scikit-learn
numpy
Werkzeug
cryptography
googletrans==4.0.0-rc1
google-cloud-translate